requirements.txt
aiohttp
async-timeout
attrs
certifi
chardet
cycler
dlib
idna
imutils
kiwisolver
matplotlib
multidict
numpy
opencv-python
PeakUtils
pyparsing
PyQt5.14.2
PyQt5-sip
pyqtgraph
python-dateutil
python-engineio
python-socketio
PyWavelets
requests
scipy
six
tk
tqdm
transforms
urllib3
websocket-client
websockets
yarl
